+++ database/migrate.sql
@@ -15,3 +15,5 @@
 \i database/migrations/013_admin_operacional.sql
 
 \i database/migrations/014_windows_iis_operations.sql
+
+\i database/migrations/015_schema_hardening.sql